+++ src/index.ts
@@ -21,20 +21,20 @@
  * limitations under the License.
  */
 
-import { Matterbridge, PlatformConfig } from 'matterbridge';
+import { PlatformConfig, PlatformMatterbridge } from 'matterbridge';
 import { AnsiLogger } from 'matterbridge/logger';
 
-import { ExampleMatterbridgeDynamicPlatform } from './platform.js';
+import { DynamicPlatformConfig, ExampleMatterbridgeDynamicPlatform } from './platform.js';
 
 /**
  * This is the standard interface for Matterbridge plugins.
  * Each plugin should export a default function that follows this signature.
  *
- * @param {Matterbridge} matterbridge - The Matterbridge instance.
+ * @param {PlatformMatterbridge} matterbridge - The Matterbridge instance.
  * @param {AnsiLogger} log - The logger instance.
  * @param {PlatformConfig} config - The platform configuration.
  * @returns {ExampleMatterbridgeDynamicPlatform} The initialized platform.
  */
-export default function initializePlugin(matterbridge: Matterbridge, log: AnsiLogger, config: PlatformConfig): ExampleMatterbridgeDynamicPlatform {
-  return new ExampleMatterbridgeDynamicPlatform(matterbridge, log, config);
+export default function initializePlugin(matterbridge: PlatformMatterbridge, log: AnsiLogger, config: PlatformConfig): ExampleMatterbridgeDynamicPlatform {
+  return new ExampleMatterbridgeDynamicPlatform(matterbridge, log, config as DynamicPlatformConfig);
 }